requirements.txt
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.3
psycopg2-binary==2.9.3
django-jsonform==2.15.0
